feat: add notification for addToPlaylist for guest
